[PIDM-307] feat: add routing for terms and privacy pages
add static page component for separation of concerns;
add error handling;
change stylesheet hrefs to absolute;
change new terms page style to match existing privacy page;
run lint --fix;
