+++ static/privacypolicy/it.html
@@ -7,7 +7,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
         <meta robots="noindex, nofollow">
         <title>Informativa sul trattamento dei dati personali - Checkout.pagopa.it</title>
-        <link href="style.css" rel="stylesheet">
+        <link href="/privacypolicy/style.css" rel="stylesheet">
     </head>
     <body>
         <h2 id="informativa-sul-trattamento-dei-dati-personali">INFORMATIVA SUL TRATTAMENTO DEI DATI PERSONALI</h2>
