+++ static/terms/it.html
@@ -7,7 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta robots="noindex, nofollow">
     <title>Termini e condizioni d'uso - Checkout.pagopa.it</title>
-    <link href="style.css" rel="stylesheet">
+    <link href="/terms/style.css" rel="stylesheet">
 </head>
 <!-- OneTrust Privacy Notice start -->
 <!-- Container in which the privacy notice will be rendered -->
